requirements-optional.txt
googledrivedownloader==0.4
onedrivedownloader==1.1.3
pytest==7.4.2
quadprog==0.1.11
setproctitle==1.3.2
wandb
deeplake<3.9
pandas
timm==0.9.8
clip @ git+https://github.com/openai/CLIP.git
scikit-learn
decorator
accelerate
transformers
kmeans_pytorch
-i https://pypi.org/simple
bitsandbytes
xitorch
dotenv
matplotlib
higher